refactoring

+++ source/Generator/tools/simplexnoise.cpp
@@ -105,7 +105,7 @@ float octave_ridgednoise_3d( const float octaves, const float persistence, const
     float maxAmplitude = 0;
 
     for( int i=0; i < octaves; i++ ) {
-        total += ((1.0f - static_cast<float>(fabs(raw_noise_3d(x * frequency, y * frequency, z * frequency)))) * 2.0f - 1.0f) * amplitude;
+        total += ((1.0 - fabs(raw_noise_3d(x * frequency, y * frequency, z * frequency))) * 2.0 - 1.0) * amplitude;
         //total += raw_noise_3d( x * frequency, y * frequency, z * frequency ) * amplitude;
 
         frequency *= 2;
@@ -197,17 +197,17 @@ float raw_noise_2d( const float x, const float y ) {
     float n0, n1, n2;
 
     // Skew the input space to determine which simplex cell we're in
-    float F2 = 0.5f * (sqrtf(3.0f) - 1.0f);
+    float F2 = 0.5 * (sqrtf(3.0) - 1.0);
     // Hairy factor for 2D
     float s = (x + y) * F2;
     int i = fastfloor( x + s );
     int j = fastfloor( y + s );
 
-    float G2 = (3.0f - sqrtf(3.0f)) / 6.0f;
-    float t = static_cast<float>(i + j) * G2;
+    float G2 = (3.0 - sqrtf(3.0)) / 6.0;
+    float t = (i + j) * G2;
     // Unskew the cell origin back to (x,y) space
-    float X0 = static_cast<float>(i)-t;
-    float Y0 = static_cast<float>(j)-t;
+    float X0 = i-t;
+    float Y0 = j-t;
     // The x,y distances from the cell origin
     float x0 = x-X0;
     float y0 = y-Y0;
@@ -221,10 +221,10 @@ float raw_noise_2d( const float x, const float y ) {
     // A step of (1,0) in (i,j) means a step of (1-c,-c) in (x,y), and
     // a step of (0,1) in (i,j) means a step of (-c,1-c) in (x,y), where
     // c = (3-sqrt(3))/6
-    float x1 = x0 - static_cast<float>(i1) + G2; // Offsets for middle corner in (x,y) unskewed coords
-    float y1 = y0 - static_cast<float>(j1) + G2;
-    float x2 = x0 - 1.0f + 2.0f * G2; // Offsets for last corner in (x,y) unskewed coords
-    float y2 = y0 - 1.0f + 2.0f * G2;
+    float x1 = x0 - i1 + G2; // Offsets for middle corner in (x,y) unskewed coords
+    float y1 = y0 - j1 + G2;
+    float x2 = x0 - 1.0 + 2.0 * G2; // Offsets for last corner in (x,y) unskewed coords
+    float y2 = y0 - 1.0 + 2.0 * G2;
 
     // Work out the hashed gradient indices of the three simplex corners
     int ii = i & 255;
@@ -234,21 +234,21 @@ float raw_noise_2d( const float x, const float y ) {
     int gi2 = perm[ii+1+perm[jj+1]] % 12;
 
     // Calculate the contribution from the three corners
-    float t0 = 0.5f - x0*x0-y0*y0;
+    float t0 = 0.5 - x0*x0-y0*y0;
     if(t0<0) n0 = 0.0;
     else {
         t0 *= t0;
         n0 = t0 * t0 * dot(grad3[gi0], x0, y0); // (x,y) of grad3 used for 2D gradient
     }
 
-    float t1 = 0.5f - x1*x1-y1*y1;
+    float t1 = 0.5 - x1*x1-y1*y1;
     if(t1<0) n1 = 0.0;
     else {
         t1 *= t1;
         n1 = t1 * t1 * dot(grad3[gi1], x1, y1);
     }
 
-    float t2 = 0.5f - x2*x2-y2*y2;
+    float t2 = 0.5 - x2*x2-y2*y2;
     if(t2<0) n2 = 0.0;
     else {
         t2 *= t2;
@@ -257,7 +257,7 @@ float raw_noise_2d( const float x, const float y ) {
 
     // Add contributions from each corner to get the final noise value.
     // The result is scaled to return values in the interval [-1,1].
-    return 70.0f * (n0 + n1 + n2);
+    return 70.0 * (n0 + n1 + n2);
 }
 
 
@@ -266,17 +266,17 @@ float raw_noise_3d( const float x, const float y, const float z ) {
     float n0, n1, n2, n3; // Noise contributions from the four corners
 
     // Skew the input space to determine which simplex cell we're in
-    float F3 = 1.0f/3.0f;
+    float F3 = 1.0/3.0;
     float s = (x+y+z)*F3; // Very nice and simple skew factor for 3D
     int i = fastfloor(x+s);
     int j = fastfloor(y+s);
     int k = fastfloor(z+s);
 
-    float G3 = 1.0f/6.0f; // Very nice and simple unskew factor, too
-    float t = static_cast<float>(i+j+k)*G3;
-    float X0 = static_cast<float>(i)-t; // Unskew the cell origin back to (x,y,z) space
-    float Y0 = static_cast<float>(j)-t;
-    float Z0 = static_cast<float>(k)-t;
+    float G3 = 1.0/6.0; // Very nice and simple unskew factor, too
+    float t = (i+j+k)*G3;
+    float X0 = i-t; // Unskew the cell origin back to (x,y,z) space
+    float Y0 = j-t;
+    float Z0 = k-t;
     float x0 = x-X0; // The x,y,z distances from the cell origin
     float y0 = y-Y0;
     float z0 = z-Z0;
@@ -301,15 +301,15 @@ float raw_noise_3d( const float x, const float y, const float z ) {
     // a step of (0,1,0) in (i,j,k) means a step of (-c,1-c,-c) in (x,y,z), and
     // a step of (0,0,1) in (i,j,k) means a step of (-c,-c,1-c) in (x,y,z), where
     // c = 1/6.
-    float x1 = x0 - static_cast<float>(i1) + G3; // Offsets for second corner in (x,y,z) coords
-    float y1 = y0 - static_cast<float>(j1) + G3;
-    float z1 = z0 - static_cast<float>(k1) + G3;
-    float x2 = x0 - static_cast<float>(i2) + 2.0f * G3; // Offsets for third corner in (x,y,z) coords
-    float y2 = y0 - static_cast<float>(j2) + 2.0f * G3;
-    float z2 = z0 - static_cast<float>(k2) + 2.0f * G3;
-    float x3 = x0 - 1.0f + 3.0f * G3; // Offsets for last corner in (x,y,z) coords
-    float y3 = y0 - 1.0f + 3.0f * G3;
-    float z3 = z0 - 1.0f + 3.0f * G3;
+    float x1 = x0 - i1 + G3; // Offsets for second corner in (x,y,z) coords
+    float y1 = y0 - j1 + G3;
+    float z1 = z0 - k1 + G3;
+    float x2 = x0 - i2 + 2.0*G3; // Offsets for third corner in (x,y,z) coords
+    float y2 = y0 - j2 + 2.0*G3;
+    float z2 = z0 - k2 + 2.0*G3;
+    float x3 = x0 - 1.0 + 3.0*G3; // Offsets for last corner in (x,y,z) coords
+    float y3 = y0 - 1.0 + 3.0*G3;
+    float z3 = z0 - 1.0 + 3.0*G3;
 
     // Work out the hashed gradient indices of the four simplex corners
     int ii = i & 255;
@@ -321,28 +321,28 @@ float raw_noise_3d( const float x, const float y, const float z ) {
     int gi3 = perm[ii+1+perm[jj+1+perm[kk+1]]] % 12;
 
     // Calculate the contribution from the four corners
-    float t0 = 0.6f - x0*x0 - y0*y0 - z0*z0;
+    float t0 = 0.6 - x0*x0 - y0*y0 - z0*z0;
     if(t0<0) n0 = 0.0;
     else {
         t0 *= t0;
         n0 = t0 * t0 * dot(grad3[gi0], x0, y0, z0);
     }
 
-    float t1 = 0.6f - x1*x1 - y1*y1 - z1*z1;
+    float t1 = 0.6 - x1*x1 - y1*y1 - z1*z1;
     if(t1<0) n1 = 0.0;
     else {
         t1 *= t1;
         n1 = t1 * t1 * dot(grad3[gi1], x1, y1, z1);
     }
 
-    float t2 = 0.6f - x2*x2 - y2*y2 - z2*z2;
+    float t2 = 0.6 - x2*x2 - y2*y2 - z2*z2;
     if(t2<0) n2 = 0.0;
     else {
         t2 *= t2;
         n2 = t2 * t2 * dot(grad3[gi2], x2, y2, z2);
     }
 
-    float t3 = 0.6f - x3*x3 - y3*y3 - z3*z3;
+    float t3 = 0.6 - x3*x3 - y3*y3 - z3*z3;
     if(t3<0) n3 = 0.0;
     else {
         t3 *= t3;
@@ -351,15 +351,15 @@ float raw_noise_3d( const float x, const float y, const float z ) {
 
     // Add contributions from each corner to get the final noise value.
     // The result is scaled to stay just inside [-1,1]
-    return 32.0f*(n0 + n1 + n2 + n3);
+    return 32.0*(n0 + n1 + n2 + n3);
 }
 
 
 // 4D raw Simplex noise
 float raw_noise_4d( const float x, const float y, const float z, const float w ) {
     // The skewing and unskewing factors are hairy again for the 4D case
-    float F4 = (sqrtf(5.0f) - 1.0f) / 4.0f;
-    float G4 = (5.0f - sqrtf(5.0f)) / 20.0f;
+    float F4 = (sqrtf(5.0)-1.0)/4.0;
+    float G4 = (5.0-sqrtf(5.0))/20.0;
     float n0, n1, n2, n3, n4; // Noise contributions from the five corners
 
     // Skew the (x,y,z,w) space to determine which cell of 24 simplices we're in
@@ -368,11 +368,11 @@ float raw_noise_4d( const float x, const float y, const float z, const float w )
     int j = fastfloor(y + s);
     int k = fastfloor(z + s);
     int l = fastfloor(w + s);
-    float t = static_cast<float>(i + j + k + l) * G4; // Factor for 4D unskewing
-    float X0 = static_cast<float>(i) - t; // Unskew the cell origin back to (x,y,z,w) space
-    float Y0 = static_cast<float>(j) - t;
-    float Z0 = static_cast<float>(k) - t;
-    float W0 = static_cast<float>(l) - t;
+    float t = (i + j + k + l) * G4; // Factor for 4D unskewing
+    float X0 = i - t; // Unskew the cell origin back to (x,y,z,w) space
+    float Y0 = j - t;
+    float Z0 = k - t;
+    float W0 = l - t;
 
     float x0 = x - X0; // The x,y,z,w distances from the cell origin
     float y0 = y - Y0;
@@ -420,22 +420,22 @@ float raw_noise_4d( const float x, const float y, const float z, const float w )
     l3 = simplex[c][3]>=1 ? 1 : 0;
     // The fifth corner has all coordinate offsets = 1, so no need to look that up.
 
-    float x1 = x0 - static_cast<float>(i1) + G4; // Offsets for second corner in (x,y,z,w) coords
-    float y1 = y0 - static_cast<float>(j1) + G4;
-    float z1 = z0 - static_cast<float>(k1) + G4;
-    float w1 = w0 - static_cast<float>(l1) + G4;
-    float x2 = x0 - static_cast<float>(i2) + 2.0f * G4; // Offsets for third corner in (x,y,z,w) coords
-    float y2 = y0 - static_cast<float>(j2) + 2.0f * G4;
-    float z2 = z0 - static_cast<float>(k2) + 2.0f * G4;
-    float w2 = w0 - static_cast<float>(l2) + 2.0f * G4;
-    float x3 = x0 - static_cast<float>(i3) + 3.0f * G4; // Offsets for fourth corner in (x,y,z,w) coords
-    float y3 = y0 - static_cast<float>(j3) + 3.0f * G4;
-    float z3 = z0 - static_cast<float>(k3) + 3.0f * G4;
-    float w3 = w0 - static_cast<float>(l3) + 3.0f * G4;
-    float x4 = x0 - 1.0f + 4.0f * G4; // Offsets for last corner in (x,y,z,w) coords
-    float y4 = y0 - 1.0f + 4.0f * G4;
-    float z4 = z0 - 1.0f + 4.0f * G4;
-    float w4 = w0 - 1.0f + 4.0f * G4;
+    float x1 = x0 - i1 + G4; // Offsets for second corner in (x,y,z,w) coords
+    float y1 = y0 - j1 + G4;
+    float z1 = z0 - k1 + G4;
+    float w1 = w0 - l1 + G4;
+    float x2 = x0 - i2 + 2.0*G4; // Offsets for third corner in (x,y,z,w) coords
+    float y2 = y0 - j2 + 2.0*G4;
+    float z2 = z0 - k2 + 2.0*G4;
+    float w2 = w0 - l2 + 2.0*G4;
+    float x3 = x0 - i3 + 3.0*G4; // Offsets for fourth corner in (x,y,z,w) coords
+    float y3 = y0 - j3 + 3.0*G4;
+    float z3 = z0 - k3 + 3.0*G4;
+    float w3 = w0 - l3 + 3.0*G4;
+    float x4 = x0 - 1.0 + 4.0*G4; // Offsets for last corner in (x,y,z,w) coords
+    float y4 = y0 - 1.0 + 4.0*G4;
+    float z4 = z0 - 1.0 + 4.0*G4;
+    float w4 = w0 - 1.0 + 4.0*G4;
 
     // Work out the hashed gradient indices of the five simplex corners
     int ii = i & 255;
@@ -449,35 +449,35 @@ float raw_noise_4d( const float x, const float y, const float z, const float w )
     int gi4 = perm[ii+1+perm[jj+1+perm[kk+1+perm[ll+1]]]] % 32;
 
     // Calculate the contribution from the five corners
-    float t0 = 0.6f - x0*x0 - y0*y0 - z0*z0 - w0*w0;
+    float t0 = 0.6 - x0*x0 - y0*y0 - z0*z0 - w0*w0;
     if(t0<0) n0 = 0.0;
     else {
         t0 *= t0;
         n0 = t0 * t0 * dot(grad4[gi0], x0, y0, z0, w0);
     }
 
-    float t1 = 0.6f - x1*x1 - y1*y1 - z1*z1 - w1*w1;
+    float t1 = 0.6 - x1*x1 - y1*y1 - z1*z1 - w1*w1;
     if(t1<0) n1 = 0.0;
     else {
         t1 *= t1;
         n1 = t1 * t1 * dot(grad4[gi1], x1, y1, z1, w1);
     }
 
-    float t2 = 0.6f - x2*x2 - y2*y2 - z2*z2 - w2*w2;
+    float t2 = 0.6 - x2*x2 - y2*y2 - z2*z2 - w2*w2;
     if(t2<0) n2 = 0.0;
     else {
         t2 *= t2;
         n2 = t2 * t2 * dot(grad4[gi2], x2, y2, z2, w2);
     }
 
-    float t3 = 0.6f - x3*x3 - y3*y3 - z3*z3 - w3*w3;
+    float t3 = 0.6 - x3*x3 - y3*y3 - z3*z3 - w3*w3;
     if(t3<0) n3 = 0.0;
     else {
         t3 *= t3;
         n3 = t3 * t3 * dot(grad4[gi3], x3, y3, z3, w3);
     }
 
-    float t4 = 0.6f - x4*x4 - y4*y4 - z4*z4 - w4*w4;
+    float t4 = 0.6 - x4*x4 - y4*y4 - z4*z4 - w4*w4;
     if(t4<0) n4 = 0.0;
     else {
         t4 *= t4;
@@ -485,12 +485,12 @@ float raw_noise_4d( const float x, const float y, const float z, const float w )
     }
 
     // Sum up and scale the result to cover the range [-1,1]
-    return 27.0f * (n0 + n1 + n2 + n3 + n4);
+    return 27.0 * (n0 + n1 + n2 + n3 + n4);
 }
 
 
-int fastfloor( const float x ) { return x > 0 ? static_cast<int>(x) : static_cast<int>(x - 1.0f); }
+int fastfloor( const float x ) { return x > 0 ? (int) x : (int) x - 1; }
 
-float dot( const int* g, const float x, const float y ) { return static_cast<float>(g[0]) * x + static_cast<float>(g[1]) * y; }
-float dot( const int* g, const float x, const float y, const float z ) { return static_cast<float>(g[0]) * x + static_cast<float>(g[1]) * y + static_cast<float>(g[2]) * z; }
-float dot( const int* g, const float x, const float y, const float z, const float w ) { return static_cast<float>(g[0]) * x + static_cast<float>(g[1]) * y + static_cast<float>(g[2]) * z + static_cast<float>(g[3]) * w; }
+float dot( const int* g, const float x, const float y ) { return g[0]*x + g[1]*y; }
+float dot( const int* g, const float x, const float y, const float z ) { return g[0]*x + g[1]*y + g[2]*z; }
+float dot( const int* g, const float x, const float y, const float z, const float w ) { return g[0]*x + g[1]*y + g[2]*z + g[3]*w; }
